requirements-render.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
google-generativeai==0.3.1
gunicorn==20.1.0
requests>=2.28.0
numpy>=1.22.0 